Refactor closing/cancelling logic for column selection.
